Implementing Api Client interface
When making a request you need to create a request from the client
now instead of creating it from a request struct.

This way we can enforce our limit and if someone decides our client
sucks they can create a new one and use that instead.

Rate Limiting and the client need to be configurable in the future.

Builds on #12